test(mcp-server): canonical schemas for attempt-scoped fixture writes refused by the D2 gate (enterprise#3045)
CI at 3e97e06c flagged three attempt-scoped fixtures whose prepared files no
longer match the canonical dataset schemas now enforced inside partition
commit attempts:

- market-datasets writeSpotPartition identity test: synthetic id/label select
  replaced with the exact revision-1 spot projection.
- market-datasets writeEnrichedTickerPartition receipt test: synthetic
  ticker/date/id/label select replaced with a shared exact revision-1
  enriched projection over src.
- parquet-spot-store-zero-guard writeFromSelect evidence test: numeric OHLC
  literals cast to DOUBLE so DuckDB does not infer DECIMAL.

Test-only; no production or gate changes.

+++ packages/mcp-server/tests/unit/parquet-spot-store-zero-guard.test.ts
@@ -125,8 +125,8 @@ describe("ParquetSpotStore writeBars zero/weekend guard", () => {
         store.writeFromSelect(
           { ticker: "SPX", date: "2025-01-07" },
           `SELECT 'SPX' AS ticker, '2025-01-07' AS date, '09:30' AS time,
-                  100.0 AS open, 101.0 AS high, 99.0 AS low, 100.5 AS close,
-                  NULL::DOUBLE AS bid, NULL::DOUBLE AS ask`,
+                  100.0::DOUBLE AS open, 101.0::DOUBLE AS high, 99.0::DOUBLE AS low,
+                  100.5::DOUBLE AS close, NULL::DOUBLE AS bid, NULL::DOUBLE AS ask`,
         ),
     );
